+++ src/app/(auth)/new-password/page.tsx
