+++ src/app/layout.tsx
@@ -3,8 +3,6 @@
 import { usePathname } from "next/navigation";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import { ThemeProvider } from "@/app/context/ThemeContext";
-import Layout from "@/components/Layout";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -14,20 +12,11 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   const pathname = usePathname();
-  const isAuthPage = pathname && pathname.startsWith("/auth");
 
   return (
     <html lang="en">
       <body className={inter.className}>
-        <ThemeProvider>
-          {isAuthPage ? (
-            children
-          ) : (
-            <Layout>
-              {children}
-            </Layout>
-          )}
-        </ThemeProvider>
+        {children}
       </body>
     </html>
   );
