--- src/app/pages/page.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function page() {
-  return (
-    <div>page</div>
-  )
-}
