+++ src/components/Sidebar.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { useState } from 'react';
 import Link from 'next/link';
 import { menuItems } from '../app/utils/menuItems';
